requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
idna==2.10
jikanpy==4.3.2
jsonfield==3.1.0
multidict==5.1.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
